requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   griffe
    #   tqdm
cryptography==46.0.3
    # via pyjwt
distro==1.9.0
    # via openai
dnspython==2.8.0
    # via email-validator
email-service==2.0.0
    # via chatbot (pyproject.toml)
email-validator==2.3.0
    # via
    #   chatbot (pyproject.toml)
    #   pydantic
fastapi==0.125.0
    # via chatbot (pyproject.toml)
griffe==1.15.0
    # via openai-agents
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   mcp
    #   openai
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jiter==0.12.0
    # via openai
jsonschema==4.25.1
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
mcp==1.24.0
    # via openai-agents
openai==2.13.0
    # via openai-agents
openai-agents==0.6.3
    # via chatbot (pyproject.toml)
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   chatbot (pyproject.toml)
    #   fastapi
    #   mcp
    #   openai
    #   openai-agents
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via mcp
pyjwt==2.10.1
    # via mcp
pypdf==6.4.2
    # via chatbot (pyproject.toml)
pypdf2==3.0.1
    # via chatbot (pyproject.toml)
python-dotenv==1.2.1
    # via pydantic-settings
python-http-client==3.3.7
    # via sendgrid
python-multipart==0.0.20
    # via mcp
pywin32==311
    # via mcp
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via openai-agents
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
sendgrid==6.4.8
    # via email-service
sniffio==1.3.1
    # via openai
sse-starlette==3.0.4
    # via mcp
starkbank-ecdsa==2.2.0
    # via sendgrid
starlette==0.50.0
    # via
    #   fastapi
    #   mcp
    #   sse-starlette
tqdm==4.67.1
    # via openai
types-requests==2.32.4.20250913
    # via openai-agents
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   mcp
    #   openai
    #   openai-agents
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
urllib3==2.6.2
    # via
    #   requests
    #   types-requests
uvicorn==0.38.0
    # via mcp
